fixed float box error

+++ src/beans_intake/forms/supplier_intake_form.py
@@ -12,11 +12,11 @@ class SupplierIntakeForm(forms.ModelForm):
     supplier_name = forms.ModelChoiceField(label="Supplier Name",
                                 queryset=Supplier.objects.all().filter(), widget=forms.Select(attrs={'class':'form-control'}))
 
-    total_box_count= forms.IntegerField(label="Number of Boxes",
+    total_box_count= forms.FloatField(label="Number of Boxes",
                                   widget=forms.NumberInput(attrs={'class': 'form-control'}),
                                   min_value = 0)
 
-    passed_float_box_count = forms.IntegerField(label="Number of Passed Float Boxes", required = False,
+    passed_float_box_count = forms.FloatField(label="Number of Passed Float Boxes", required = False,
                                   widget=forms.NumberInput(attrs={'class': 'form-control'}),
                                   min_value = 0)
